refactor: dynamics library
