+++ pyproject.toml
@@ -206,7 +206,7 @@
     ]
     "noxfile.py" = ["ERA001", "T20"]
     "__init__.py" = ["F403"]
-    "__init__.pyi" = ["F401"]
+    "__init__.pyi" = ["F401", "F403"]
 
 
   [tool.ruff.lint.isort]
