+++ src/galax/dynamics/__init__.py
@@ -1,13 +1,6 @@
-"""galax: Galactic Dynamix in Jax."""
+""":mod:`galax.dynamics`."""
 
-from . import _base, _core, _orbit, mockstream
-from ._base import *
-from ._core import *
-from ._orbit import *
-from .mockstream import *
+from . import _dynamics
+from ._dynamics import *
 
-__all__: list[str] = []
-__all__ += _base.__all__
-__all__ += _core.__all__
-__all__ += _orbit.__all__
-__all__ += mockstream.__all__
+__all__ = _dynamics.__all__
