+++ src/galax/dynamics/_dynamics/__init__.py
@@ -1,0 +1,17 @@
+"""galax: Galactic Dynamix in Jax."""
+
+from . import base, core, mockstream, orbit
+from .base import *
+from .core import *
+from .mockstream import *
+from .orbit import *
+
+__all__: list[str] = ["mockstream"]
+__all__ += base.__all__
+__all__ += core.__all__
+__all__ += orbit.__all__
+__all__ += mockstream.__all__
+
+
+# Cleanup
+del base, core, orbit
