+++ src/galax/dynamics/_dynamics/core.py
@@ -13,7 +13,7 @@ from galax.typing import BatchFloatScalar, BroadBatchFloatScalar, BroadBatchVec3
 from galax.utils._shape import batched_shape, expand_batch_dims
 from galax.utils.dataclasses import converter_float_array
 
-from ._base import AbstractPhaseSpacePosition
+from .base import AbstractPhaseSpacePosition
 
 if TYPE_CHECKING:
     from galax.potential._potential.base import AbstractPotentialBase
