+++ src/galax/dynamics/_dynamics/mockstream/df/__init__.py
@@ -5,6 +5,9 @@ from . import base, fardal
 from .base import *
 from .fardal import *
 
-__all__ = []
+__all__: list[str] = []
 __all__ += base.__all__
 __all__ += fardal.__all__
+
+# Cleanup
+del base, fardal
