+++ src/galax/dynamics/_dynamics/mockstream/df/fardal.py
