--- src/galax/dynamics/mockstream/__init__.py
@@ -1,11 +1,0 @@
-"""galax: Galactic Dynamix in Jax."""
-
-from . import _core, _df, _mockstream_generator
-from ._core import *
-from ._df import *
-from ._mockstream_generator import *
-
-__all__: list[str] = []
-__all__ += _df.__all__
-__all__ += _core.__all__
-__all__ += _mockstream_generator.__all__
