+++ tests/smoke/dynamics/mockstream/test_package.py
@@ -1,12 +1,11 @@
 """Testing :mod:`galax.dynamics.mockstream` module."""
 
 from galax.dynamics import mockstream
+from galax.dynamics._dynamics.mockstream import core, df, mockstream_generator
 
 
 def test_all() -> None:
     """Test the `galax.dynamics.mockstream` API."""
     assert set(mockstream.__all__) == set(
-        mockstream._df.__all__
-        + mockstream._core.__all__
-        + mockstream._mockstream_generator.__all__
+        df.__all__ + core.__all__ + mockstream_generator.__all__
     )
