+++ tests/smoke/dynamics/test_package.py
@@ -1,10 +1,15 @@
 """Testing :mod:`galax.dynamics` module."""
 
 import galax.dynamics as gd
+from galax.dynamics._dynamics import base, core, mockstream, orbit
 
 
 def test_all() -> None:
     """Test the `galax.potential` API."""
-    assert set(gd.__all__) == set(
-        gd._base.__all__ + gd._core.__all__ + gd._orbit.__all__ + gd.mockstream.__all__
-    )
+    assert set(gd.__all__) == {
+        "mockstream",
+        *base.__all__,
+        *core.__all__,
+        *orbit.__all__,
+        *mockstream.__all__,
+    }
